+++ apps/example/src/test.view.ts
@@ -1,42 +1,28 @@
 import { View, renderToString } from "@dlightjs/dlight"
 import { css, div } from "@dlightjs/easy-css"
-import { button, htmlTag, Prop, required, SubView } from "@dlightjs/types"
+import { button, Env, htmlTag, Prop, required, SubView } from "@dlightjs/types"
 import { HStack, Route, RouterSpace, VStack } from "@dlightjs/components"
 import { MarkitView, addBlockRule } from "@dlightjs/markit"
 
 class NNN extends View {
+  @Env path: any
   Body() {
-    _(this._$children)
+    div(this.path)
   }
 }
 
 class TestView extends View {
   count = 5
 
-  onclick() {
-    this.count--
-    console.log(this.count)
-  }
-
-  @SubView
-  JJ() {
-    div("jhh")
-      .color("blue")
-  }
-
-  jj = { no: "44" }
-
   Body() {
-    button("-")
-      .onclick(() => {
-        this.jj = undefined
-      })
-    button("+")
-      .onclick(() => {
-        this.jj = { no: "55" }
-      })
-    if (this.jj) {
-      div(this.jj.no)
+    "jjj"
+    RouterSpace()
+    {
+      Route("hello")
+      {
+        div("hello")
+        NNN()
+      }
     }
   }
 }