+++ apps/test/CHANGELOG.md
@@ -1,5 +1,12 @@
 # dlight-test
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/components@0.9.29
+
 ## 0.0.3
 
 ### Patch Changes