+++ packages/ecosystem/components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/components
 
+## 0.9.29
+
+### Patch Changes
+
+- RouterSpace add currUrl
+
 ## 0.9.28
 
 ### Patch Changes